Move some tests to an init specific test file
