+++ packages/cli/lib/sync.unit.cli-test.ts
@@ -5,11 +5,10 @@ import fs from 'node:fs';
 import yaml from 'js-yaml';
 import stripAnsi from 'strip-ansi';
 import { generate } from './cli.js';
-import { exampleSyncName } from './constants.js';
 import { init } from './services/init.service.js';
 import { compileAllFiles, compileSingleFile, getFileToCompile } from './services/compile.service.js';
 import parserService from './services/parser.service.js';
-import { copyDirectoryAndContents, removeVersion, fixturesPath, getTestDirectory } from './tests/helpers.js';
+import { copyDirectoryAndContents, fixturesPath, getTestDirectory } from './tests/helpers.js';
 import { parse } from './services/config.service.js';
 import { directoryMigration, endpointMigration } from './services/migration.service.js';
 
@@ -19,35 +18,6 @@ describe('generate function tests', () => {
 
     afterEach(() => {
         consoleMock.mockReset();
-    });
-
-    it('should init the expected files in the nango-integrations directory', async () => {
-        const dir = await getTestDirectory('init');
-        init({ absolutePath: path.resolve(dir), debug: false });
-        expect(fs.existsSync(join(dir, '.nango'))).toBe(true);
-        expect(fs.existsSync(join(dir, '.nango/.gitkeep'))).toBe(true);
-        expect(fs.existsSync(join(dir, `demo-github-integration/syncs/${exampleSyncName}.ts`))).toBe(true);
-        expect(fs.existsSync(join(dir, '.env'))).toBe(true);
-        expect(fs.existsSync(join(dir, 'nango.yaml'))).toBe(true);
-        expect(fs.existsSync(join(dir, 'models.ts'))).toBe(true);
-        expect(removeVersion(fs.readFileSync(join(dir, '.nango/schema.ts')).toString())).toMatchSnapshot();
-        expect(removeVersion(fs.readFileSync(join(dir, '.nango/schema.json')).toString())).toMatchSnapshot();
-    });
-
-    it('should not overwrite existing integration files', async () => {
-        const dir = await getTestDirectory('overwrite');
-        init({ absolutePath: dir, debug: false });
-        await fs.promises.writeFile(join(dir, `${exampleSyncName}.ts`), 'dummy fake content', 'utf8');
-
-        const dummyContent = 'This is dummy content. Do not overwrite!';
-        const exampleFilePath = join(dir, `${exampleSyncName}.ts`);
-        await fs.promises.writeFile(exampleFilePath, dummyContent, 'utf8');
-
-        init({ absolutePath: dir });
-
-        expect(fs.existsSync(exampleFilePath)).toBe(true);
-        const fileContentAfterInit = await fs.promises.readFile(exampleFilePath, 'utf8');
-        expect(fileContentAfterInit).toBe(dummyContent);
     });
 
     it('should generate a different sync correctly', async () => {
